Organize implementations of constrained optimization
